Ensuring out/cout renders a text integer

+++ src/Menu.Function.cpp
@@ -6,7 +6,7 @@ namespace FactUtilEmbedded
   void renderParameterError(uint8_t expectedParameterCount, uint8_t count)
   {
     cout << endl;
-    cout << F("Expected ") << expectedParameterCount;
-    cout << F(" parameters but instead got ") << count;
+    cout << F("Expected ") << (uint16_t)expectedParameterCount;
+    cout << F(" parameters but instead got ") << (uint16_t)count;
   }
 }
